[Refactor] dependencies에서 싱글톤 패턴 정의
